Factor out time keeping code into its own module: mono_time.c.
It turns out, `unix_time` is also monotonic, and is used as such, so I've
renamed the new functions to `mono_time_*`.

2018-07-08:
```
00:01 <@irungentoo> the idea used to be that the unix_time() function
  could go backward in time but I think I might have started using it like
  if it could not after I changed it so that it would never go back in time
```

+++ toxcore/Messenger.c
@@ -34,6 +34,7 @@
 #include <time.h>
 
 #include "logger.h"
+#include "mono_time.h"
 #include "network.h"
 #include "state.h"
 #include "util.h"
@@ -1764,8 +1765,8 @@ static int handle_filecontrol(Messenger *m, int32_t friendnumber, uint8_t receiv
 
             if (position >= ft->size) {
                 LOGGER_DEBUG(m->log,
-                             "file control (friend %d, file %d): seek position %lld exceeds file size %lld",
-                             friendnumber, filenumber, (unsigned long long)position, (unsigned long long)ft->size);
+                             "file control (friend %d, file %d): seek position %ld exceeds file size %ld",
+                             friendnumber, filenumber, (unsigned long)position, (unsigned long)ft->size);
                 return -1;
             }
